Get rid of LitePyTorchAndroid
