+++ android/pytorch_android/build.gradle
@@ -41,7 +41,6 @@ android {
                 println 'Build pytorch_jni'
                 exclude 'org/pytorch/LiteModuleLoader.java'
                 exclude 'org/pytorch/LiteNativePeer.java'
-                exclude 'org/pytorch/LitePyTorchAndroid.java'
               } else {
                 println 'Build pytorch_jni_lite'
               }
