+++ android/pytorch_android/src/main/cpp/pytorch_jni_jit.cpp
@@ -17,11 +17,6 @@
 #include <android/asset_manager.h>
 #include <android/asset_manager_jni.h>
 #include <android/log.h>
-
-#ifndef USE_PTHREADPOOL
-#define USE_PTHREADPOOL
-#endif /* USE_PTHREADPOOL */
-#include <caffe2/utils/threadpool/pthreadpool-cpp.h>
 #endif
 
 namespace pytorch_jni {
@@ -239,34 +234,6 @@ class PytorchJni : public facebook::jni::HybridClass<PytorchJni> {
         methodName.c_str());
   }
 };
-
-#if defined(__ANDROID__)
-class PyTorchAndroidJni : public facebook::jni::JavaClass<PyTorchAndroidJni> {
- public:
-  constexpr static auto kJavaDescriptor = "Lorg/pytorch/PyTorchAndroid;";
-
-  static void registerNatives() {
-    javaClassStatic()->registerNatives({
-        makeNativeMethod(
-            "nativeSetNumThreads", PyTorchAndroidJni::setNumThreads),
-    });
-  }
-
-  static void setNumThreads(facebook::jni::alias_ref<jclass>, jint numThreads) {
-    caffe2::pthreadpool()->set_thread_count(numThreads);
-  }
-};
-#endif
-
-void common_registerNatives() {
-  static const int once = []() {
-#if defined(__ANDROID__)
-    pytorch_jni::PyTorchAndroidJni::registerNatives();
-#endif
-    return 0;
-  }();
-  ((void)once);
-}
 
 } // namespace pytorch_jni
